+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Tucan.scatter(:iris, "petal_width", "petal_length")
 
 - Flipped `:orient` semantics for `Tucan.bar/4`
 - Rename `:group` option to `:group_by` in `Tucan.stripplot/3` for consistency.
+- Rename `:groupby` option to `:group_by` in `Tucan.density/3` for consistency.
 
 ### Deprecated